Update of pct.bib, update of pCT_Data_Organization_Format.pdf + .tex files, and addition of overall pCT_Documentation.pdf + .tex files
Update of pct.bib:

The pct.bib file has been updated for subsequent usage by the pCT collaboration, with a description of the naming/formatting scheme used in defining references and instructions for how users should define new publications they add to it. The goal is for people to contribute to growing this file to include referencing information for every publication that is relevant to pCT publications so correctly formatted references are available for usage in writing new publications and these all appear the same in each publication without requiring users to manually define or search for the BibTeX citation info every time. The currently defined references contained in this file have been modified to conform with the naming/formatting scheme outlined in the document, but there are still some references remaining that still need such modifications, so these will be addressed in the near future.

Update of pCT_Data_Organization_Format.pdf + .tex files:

The PDF defining the standardized pCT data naming and organizational scheme was updated to reflect some of the changes to the code/data storage on Baylor's Kodiak and Tardis clusters. There are some additional changes that have recently been made to Kodiak/Tardis that may require some modifications to be made to this document and if so, these will be made in the near future and included in a subsequent commit. In any case, the information contained in this document should be sufficient for users to understand and begin interacting with Kodiak/Tardis, particularly in terms of the code/data naming and organizational standard as this remains the same.

Addition of overall pCT_Documentation.pdf + .tex files:

A new PDF and the .tex files used to generate it were added to a new /doc folder to provide documentation with a general overview of the overall topic and status of pCT research, such as the scanner system and software/hardware, the preprocessing and reconstruction programs, and other relevant information. This is in its initial stages of development, so none of the sections of the document provide a complete overview, but the hope is that by making it accessible by everyone, that those with expertise in each of the topics will contribute to the writing of its documentation so it is written more accurately and thoroughly and is completed more quickly.

The goal is for this document to be completed ASAP with a complete overview of the project status which is then periodically updated to include all recent development  such that it becomes a general resource for people, particularly those joining the collaboration and beginning their own work on it, to become familiar with pCT and the current state of its research. The document is broken into separate "Part" sections and the content for these are provided in separate .tex files such that people can work on the different parts independently at the same time without interfering with each other and creating file conflicts.

NOTE: Each of the documentation PDFs are generated in LaTeX using a special package developed by the R2Labs research group at Baylor, which is currently only available from the BlakeSchultze/LaTeX-Packages repository on GitHub. This will be added to CTAN and the Miktex/TexLive package distributions, but it will also be added directly to these repositories in the near future so users may compile the .tex files and generate the PDFs themselves; this package was specifically designed to be portable and allows for the option for its files to be added into and used from any local directory instead of requiring it to be added to the LaTeX distribution directory (as with other packages). In the mean time, although people will not be able to compile and generate the full PDF themselves unless they acquire the library from GitHub, this does not prevent users from working on developing the constituent .tex files since the package still supports all basic .tex command/environment usage. Users may also refer to several .tex files with the documentation that show the usage and resulting output appearance of many (if not all) of the new commands/environments provided by the package that will be useful in generating the documentation PDFs (see e.g. pCT_Data_Organization_Format.tex/PDF and its constituent files for an example of a complete document generated using the package's commands/environments, as these are the same ones that will be used in generating the other documentation PDFs).